Refactor data test packages
